I no longer care about "DesignForExtension"
It's been a pain in the ass, and it is unobtainable.
AgentSettings can't be final, or have final methods.  Java Persistence
doesn't work if it is either of those things.
A final method in pircbot is also currently blocking WebIRC support.
That hints to me that perhaps making everything not able to be
overridden is a bad thing.