Moved IRCAgent initialization into it's own thread
Profileing revealed that onPlayerJoin was very slow. The culprit turned
out to be both IRCAgent intialization and the intitial call to
connect().  This change moves these 2 calls into a new thread.

This change also changes the bot list from a trove intmap to a java
concurrent hashmap.  This drops the need for trove, so it was removed
completely